Add error boundaries to screen components
