+++ package.json
@@ -26,6 +26,7 @@
     "react-copy-to-clipboard": "^5.0.1",
     "react-device-detect": "^1.15.0",
     "react-dom": "^17.0.1",
+    "react-error-boundary": "^3.1.0",
     "react-helmet": "^6.1.0",
     "react-modal": "^3.12.1",
     "react-notify-toast": "^0.5.0",
