+++ src/components/UI/ErrorSection.js
@@ -1,17 +1,16 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-function ErrorSection(props) {
-  const errorMessage = props.message || "General error";
+function ErrorSection({ children }) {
   return (
-    <div data-testid="error-section" className="error-section">
-      Error: {errorMessage}
+    <div role="alert" data-testid="error-section" className="error-section">
+      {children}
     </div>
   );
 }
 
 ErrorSection.propTypes = {
-  message: PropTypes.string,
+  children: PropTypes.node,
 };
 
 export default ErrorSection;
