+++ src/registerFaIcons.js
@@ -10,6 +10,7 @@ import {
   faCode,
   faCopy,
   faDownload,
+  faExclamationTriangle,
   faExpand,
   faFileCode,
   faImage,
@@ -32,6 +33,7 @@ export default function registerIcons() {
     faCode,
     faCopy,
     faDownload,
+    faExclamationTriangle,
     faExpand,
     faFileCode,
     faImage,
